refactor(api): build get_my_proposals through a services module
The endpoint had grown a second query and its grouping helper inline, which is
not where the events package keeps that work. The query moves to services.py
behind a thin whitelisted wrapper, matching buzz/api/events.

The rows also carry event.end_date now. The timeline splits upcoming from past
on the date an event actually ends, and without the column a multi-day event
still running today filed itself under past.

The three single-column tests collapse into one: they each re-asserted a
joined field the shape test already covers.
